Allow multiple add_remark_slides() in one directory.
I tend to organize my presentations by **event**, rather than topic.
So for conference *X*, I might have several presentations, and they'll
all go in the *X* directory. Each add_remark_slides() writes a
`gen_slides.cmake` to do the actual work .. and so having multiple
calls to that function within a single directory (single CMakeLists.txt)
overwrites one with the other. Then you get weird errors on running
*make* because it prints *Generating talk-sunday* and then errors
out because inside `gen_slides.cmake` it does the work for talk-saturday.

This patch refactors the name of the output file to a function-local
variable `${gen_slides}` and introduces the NAME of the presentation
(given as argument to `add_remark_slides()`) into the filename,
so that they don't overwrite each other.